fix(refactory): repo online and repo location

+++ simple-geo-nearby-users/api/handlers/dtos/create-location.go
@@ -2,6 +2,6 @@ package handlers
 
 type CreateLocationRequestBody struct {
 	Username  string  `json:"username" binding:"required"`
-	Latitude  float32 `json:"latitude" binding:"required"`
-	Longitude float32 `json:"longitude" binding:"required"`
+	Latitude  float64 `json:"latitude" binding:"required"`
+	Longitude float64 `json:"longitude" binding:"required"`
 }
